+++ lib/burner/payload.rb
@@ -7,6 +7,8 @@
 # LICENSE file in the root directory of this source tree.
 #
 
+require_relative 'data'
+
 module Burner
   # The input for all Job#perform methods.  The main notion of this object is its 'registers'
   # attribute.  This registers attribute is a key-indifferent hash, accessible on Payload using
@@ -24,10 +26,15 @@ module Burner
   # The 'time' attribute is important in that it should for the replaying of pipelines and jobs.
   # Instead of having job's utilizing Time.now, Date.today, etc... they should rather opt to
   # use this value instead.
+  #
+  # The notion of params are somewhat conflated with registers here.  Both are hashes and both
+  # serve as data stores. The difference is registers are really meant to be the shared-data
+  # repository across jobs, while params are, more or less, the input into the _pipeline_.
+  # It is debatable if mutation of the params should be allowed but the design decision was made
+  # early on to allow for params to be mutable albeit with registers being the preferred mutable
+  # store.
   class Payload
-    attr_reader :params,
-                :registers,
-                :side_effects,
+    attr_reader :side_effects,
                 :time
 
     def initialize(
@@ -36,12 +43,32 @@ module Burner
       side_effects: [],
       time: Time.now.utc
     )
-      @params       = params || {}
-      @registers    = {}
+      @params       = Data.new(params)
+      @registers    = Data.new(registers)
       @side_effects = side_effects || []
       @time         = time || Time.now.utc
+    end
 
-      add_registers(registers)
+    # Backwards compatibility.  This allows for control over the underlying data structure
+    # while still maintaining the same public API as before.
+    def params
+      @params.to_h
+    end
+
+    # Backwards compatibility.  This allows for control over the underlying data structure
+    # while still maintaining the same public API as before.
+    def registers
+      @registers.to_h
+    end
+
+    # Law of Demeter: While params is an accessible hash, it is preferred that the
+    # public class methods are used.
+    def param(key)
+      @params[key]
+    end
+
+    def update_param(key, value)
+      tap { @params[key] = value }
     end
 
     # Add a side effect of a job.  This helps to keep track of things jobs do outside of its
@@ -52,12 +79,12 @@ module Burner
 
     # Set a register's value.
     def []=(key, value)
-      set(key, value)
+      @registers[key] = value
     end
 
     # Retrieve a register's value.
     def [](key)
-      registers[key.to_s]
+      @registers[key]
     end
 
     # Set halt_pipeline to true.  This will indicate to the pipeline to stop all
@@ -71,14 +98,11 @@ module Burner
       @halt_pipeline || false
     end
 
-    private
+    def params_and_registers_hash
+      registers_hash = @registers.transform_keys { |key| "__#{key}_register" }
+      params_hash    = @params.to_h
 
-    def set(key, value)
-      registers[key.to_s] = value
-    end
-
-    def add_registers(registers)
-      (registers || {}).each { |k, v| set(k, v) }
+      params_hash.merge(registers_hash)
     end
   end
 end